+++ src/lib/domain/qualifications.ts
